refactor(checker): split generator recovery helper
