+++ crates/tsz-checker/src/checkers/mod.rs
@@ -12,6 +12,7 @@ pub mod iterable_checker;
 pub mod jsx;
 pub mod parameter_checker;
 pub mod promise_checker;
+mod promise_checker_generator;
 mod promise_checker_object_normalization;
 pub mod property_checker;
 pub mod signature_builder;
